qrouting: Rework/Fix bogus gwlist sorting logic
The sorting logic was incorrect, as the carriers were never actually
sorted within the gwlist array -- their positions were always fixed.
Only groups of gateways in-between these carriers were getting sorted
by their qrouting score.  For example, assuming the numbers represent
qrouting-assigned scores:

this rule gwlist: "4, 2, 1, CR1=0.5, 1, 0" would get sorted as:

"1, 2, 4, CR1=0.5, 0, 1", whereas the correct sort would have been:

"0, CR1=0.5, 1, 1, 2, 4"

+++ modules/qrouting/qr_stats.h
@@ -96,7 +96,7 @@ typedef struct qr_gw {
 	qr_stats_t current_interval; /* the current interval */
 	qr_stats_t summed_stats; /* the sum of the @lru_interval list */
 	char state;
-	int score; /* the score of the gateway (based on thresholds) */
+	double score; /* score of the gateway, based on weights & thresholds */
 	rw_lock_t *ref_lock; /* lock for protecting the overall statistics (history) */
 	gen_lock_t *acc_lock; /* lock for protecting the current interval */
 } qr_gw_t;
